Allow analysis of remote repos
This code changes the input params to let run
RepoFS on remote repositories. The tool requires
an URI which targets the repo to process, two
optional params can be used to define the
local paths of the cloned repo and file system.

The new params are:
- Uri, which points to the local/remote repo
- Git-path (optional), which points where the repo
  will be cloned locally. If null, the default
  root location is ~/.repofs/repositories
- Mount-path (optional), which points where the FS will
  be mounted. If null, the default root location is
  ~/.repofs/mounts

An example of RepoFS call from cmd is:
```
repofs https://github.com/org/repo
```

When the tool is executed, the target repo is cloned
only if it doesn't exist, otherwise pygit2 is used
to pull all branches. The corresponding code added in
the main() relies on https://github.com/MichaelBoselowitz/pygit2-examples/blob/master/examples.py#L54.

The logic included in __main__ has been moved to /bin/repofs
for refactor purposes.

The requirements.txt is included to test the tool as
follows:
```
pip install -r requirements.txt -e .
```